+++ src/Typin/Typin.Tests/Data/Modes/Invalid/InvalidAbstractCustomMode.cs
@@ -15,7 +15,7 @@
             _console = console;
         }
 
-        public async ValueTask<int> ExecuteAsync(IReadOnlyList<string> commandLineArguments, ICliCommandExecutor executor)
+        public async ValueTask<int> ExecuteAsync(IEnumerable<string> commandLineArguments, ICliCommandExecutor executor)
         {
             await _console.Output.WriteLineAsync(nameof(InvalidCustomMode));