+++ src/Typin/Typin/Utilities/CommandLineSplitter.cs
@@ -65,7 +65,9 @@
                 }
                 else if (chr == ' ' && !quoted)
                 {
-                    if (started) yield return result.ToString();
+                    if (started)
+                        yield return result.ToString();
+
                     result.Clear();
                     started = false;
                 }